+++ Packages/ai.muna.muna/Runtime/Beta/OpenAI/Types.cs
@@ -1,0 +1,10 @@
+/* 
+*   Muna
+*   Copyright © 2025 NatML Inc. All rights reserved.
+*/
+
+#nullable enable
+
+namespace Muna.Beta.OpenAI {
+
+}